src/requirements.txt
asyncpg==0.30.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
Django==5.2.8
djangorestframework==3.16.1
exceptiongroup==1.3.0
fastapi==0.121.2
idna==3.11
psycopg2==2.9.11
psycopg2-binary==2.9.10
pydantic==2.12.4
pydantic_core==2.41.5
python-dotenv == 1.2.1
sniffio==1.3.1
sqlparse==0.5.3
starlette==0.49.3
sqlalchemy == 2.0.44
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0.